Extract abstract streaming data calculator
